update for new lingon logger api and check if build folder exists

+++ lib/deploy.sh
@@ -5,6 +5,11 @@ BRANCH=$2
 REMOTE_URL=$3
 REMOTE="deploy"
 
+if [ ! -d "$CWD" ]; then
+  echo "$CWD doesn't seem to exist. Make sure you run 'lingon build' before deploying"
+  exit 1
+fi
+
 cd $CWD
 rm -rf .git
 git init .
